feat: docs moved to a separate package
