+++ app/index.html
@@ -1,52 +1,33 @@
 <!DOCTYPE html>
 <html lang="en">
-	<head>
-		<title>@gun-vue demo app</title>
 
-		<link rel="manifest" href="/manifest.json" />
+<head>
+	<title>@gun-vue demo app</title>
 
-		<meta
-			name="description"
-			content="Gun-Vue – is where the peer-to-peer javascript database Gun meets the reactivity system of Vue 3."
-		/>
-		<meta name="theme-color" content="#ff0088" />
-		<meta charset="UTF-8" />
-		<link
-			rel="icon"
-			href="/media/gun-vue-logo.svg"
-			sizes="any"
-			type="image/svg+xml"
-		/>
-		<link rel="mask-icon" href="/media/gun-vue-logo.svg" color="#FFFFFF" />
-		<meta
-			name="viewport"
-			content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no"
-		/>
-		<meta name="twitter:card" content="summary_large_image" />
-		<meta name="twitter:site" content="@defucc" />
-		<meta name="twitter:creator" content="@defucc" />
-		<meta
-			name="twitter:title"
-			content="Gun-Vue – is where the peer-to-peer javascript database Gun meets the reactivity system of Vue 3. "
-		/>
-		<meta
-			name="twitter:description"
-			content="Pluggable components for UX designers. Easily importable composables for UI developers. Relay peer for p2p network enthusiasts. Demos and docs for beginners. Open source for everyone! ✊"
-		/>
-		<meta
-			name="twitter:image"
-			content="https://gun-vue.js.org/media/gun-vue-card.png"
-		/>
-		<script
-			async
-			defer
-			data-website-id="5b80bbe4-5a00-4c06-8afe-8776c6848c87"
-			src="https://stats.defucc.me/umami.js"
-		></script>
-	</head>
+	<link rel="manifest" href="/manifest.json" />
 
-	<body>
-		<div id="app"></div>
-		<script type="module" src="./src/main.ts"></script>
-	</body>
+	<meta name="description"
+		content="Gun-Vue – is where the peer-to-peer javascript database Gun meets the reactivity system of Vue 3." />
+	<meta name="theme-color" content="#ff0088" />
+	<meta charset="UTF-8" />
+	<link rel="icon" href="/media/gun-vue-logo.svg" sizes="any" type="image/svg+xml" />
+	<link rel="mask-icon" href="/media/gun-vue-logo.svg" color="#FFFFFF" />
+	<meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
+	<meta name="twitter:card" content="summary_large_image" />
+	<meta name="twitter:site" content="@defucc" />
+	<meta name="twitter:creator" content="@defucc" />
+	<meta name="twitter:title"
+		content="Gun-Vue – is where the peer-to-peer javascript database Gun meets the reactivity system of Vue 3. " />
+	<meta name="twitter:description"
+		content="Pluggable components for UX designers. Easily importable composables for UI developers. Relay peer for p2p network enthusiasts. Demos and docs for beginners. Open source for everyone! ✊" />
+	<meta name="twitter:image" content="https://gun-vue.js.org/media/gun-vue-card.png" />
+	<script async defer data-website-id="5b80bbe4-5a00-4c06-8afe-8776c6848c87"
+		src="https://stats.defucc.me/umami.js"></script>
+</head>
+
+<body>
+	<div id="app"></div>
+	<script type="module" src="../src/main.ts"></script>
+</body>
+
 </html>
