+++ app/tsconfig.json
@@ -25,6 +25,6 @@
 		}
 	},
 	"include": [
-		"./src"
+		"../src"
 	]
 }
